refactor: move not biz code to utils
