+++ src/pages/api/chatgpt/user.ts
@@ -1,9 +1,7 @@
-import { createCipheriv, createDecipheriv, randomBytes, createHash } from "node:crypto";
-const hasher = createHash("sha256");
-
-import { NextApiHandler, NextApiRequest, NextApiResponse } from "next";
+import { NextApiHandler } from "next";
 import { SITE_USER_COOKIE } from "@/configs/constants";
-import { createUser, getUserByKeyHashed, isValidUser } from "@/storage/planetscale";
+import { createUser, isValidUser } from "@/storage/planetscale";
+import { encryptedKey, hashedKey } from "@/uitls/crypto.util";
 
 // type Request = {
 //   action: "login" | "logout";
@@ -15,35 +13,7 @@ type Response = {
   error?: string;
 };
 
-function genIV() {
-  return new Buffer(randomBytes(16));
-}
-
-function encrypt(data: string, secret: string, iv: Buffer) {
-  const cipher = createCipheriv("aes-256-cbc", secret, iv);
-  let encrypted = cipher.update(data, "utf8", "hex");
-  encrypted += cipher.final("hex");
-  return encrypted;
-}
-
-export function decrypt(encrypted: string, secret: string, iv: string) {
-  const ivBuffer = new Buffer(iv, "hex");
-  const decipher = createDecipheriv("aes-256-cbc", secret, ivBuffer);
-  let decrypted = decipher.update(encrypted, "hex", "utf8");
-  decrypted += decipher.final("utf8");
-  return decrypted;
-}
-
-export const secret = process.env["ENC_KEY"];
-
 const handler: NextApiHandler = async (req, res) => {
-  if (!secret) {
-    res.status(500).json({
-      error: "No secret key env in the server.",
-    });
-    return;
-  }
-
   if (!(req.method === "POST" && req.body)) {
     res.status(404).json({ error: "Not found" });
     return;
@@ -60,12 +30,10 @@ const handler: NextApiHandler = async (req, res) => {
   switch (action) {
     case "login":
       if (key) {
-        const key_hashed = hasher.copy().update(key).digest().toString("hex");
+        const key_hashed = hashedKey(key);
 
         if (!(await isValidUser(key_hashed))) {
-          const iv = genIV();
-          const key_encrypted = encrypt(key, secret, iv);
-
+          const { iv, key_encrypted } = encryptedKey(key);
           await createUser({
             iv: iv.toString("hex"),
             key_hashed,
@@ -90,24 +58,3 @@ const handler: NextApiHandler = async (req, res) => {
   }
 };
 export default handler;
-
-export type User = Awaited<ReturnType<typeof getUserByKeyHashed>>;
-export async function getUser(req: NextApiRequest, res: NextApiResponse): Promise<User | null> {
-  const keyHashed = req.cookies[SITE_USER_COOKIE];
-  if (!keyHashed) {
-    res.status(400).json({ error: "You're not logged in yet!" });
-    return null;
-  }
-
-  const user = await getUserByKeyHashed(keyHashed);
-  if (!user) {
-    kickOutUser(res);
-    res.status(400).json({ error: "Your login session has been expired!" });
-    return null;
-  }
-  return user;
-}
-
-export function kickOutUser(res: NextApiResponse) {
-  res.setHeader("Set-Cookie", `${SITE_USER_COOKIE}=; Max-Age=0; HttpOnly; Path=/;`);
-}
